Moved preprocessing utility classes to SamplesConsole
