+++ ISAAR.MSolve.SamplesConsole/FEM/Cantilever2DPreprocessor.cs
@@ -10,7 +10,7 @@ using ISAAR.MSolve.Materials;
 using ISAAR.MSolve.Preprocessor.Meshes;
 using ISAAR.MSolve.Preprocessor.Meshes.Custom;
 using ISAAR.MSolve.Preprocessor.Meshes.GMSH;
-using ISAAR.MSolve.Preprocessor.UI;
+using ISAAR.MSolve.SamplesConsole.Preprocessing;
 
 namespace ISAAR.MSolve.SamplesConsole.FEM
 {
